doc(solver): Clean up and document the justified-propagator interface.
