+++ solver/src/reasoners/cp/mul.rs
@@ -297,7 +297,7 @@ mod test {
     use rand::{Rng, SeedableRng, rngs::SmallRng};
 
     use super::*;
-    use crate::{core::*, reasoners::cp::test::utils::test_explanations};
+    use crate::{core::*, reasoners::cp::propagator::test::utils::test_explanations};
 
     // === Assertions ===
